+++ .old/lib2/utils/settings.dart
@@ -1,0 +1,36 @@
+// ===============================================================
+// Url de el Servidor para hacer peticiones
+// ===============================================================
+final String url = 'https://another-backend.herokuapp.com/palabras/azar';
+// ===============================================================
+// Url de el Servidor para enviar comentarios
+// ===============================================================
+final String urlSendFeedback = 'https://another-backend.herokuapp.com/send/feedback';
+// ===============================================================
+// Url de Flutter
+// ===============================================================
+final String flutterUrl = 'https://flutter.io';
+// ===============================================================
+// Nombre de la aplicación
+// ===============================================================
+final String appname = 'American English Words';
+// ===============================================================
+// Nombres de la aplicación
+// ===============================================================
+final String savedSection = 'Palabras Guardadas';
+final String settingsSection = 'Configuración';
+final String infoSection = 'Información';
+final String buttonName = 'Escuchar pronunciación';
+// ===============================================================
+// Success messages
+// ===============================================================
+final String onSuccessMessage = 'Palabra guardada correctamente';
+final String onSuccessDeletingWord = 'Palabra eliminada correctamente';
+// ===============================================================
+// Error messages
+// ===============================================================
+final String onErrorMessage = 'Hubo un problema al guardar los datos';
+final String onErrorDeletingMessage = 'Hubo un problema al eliminar los datos';
+final String onNoSavedWords = '¡Parece que aún no has guardado ninguna palabra!';
+final String onSavedWordsError = 'Hubo un error :(';
+final String onServerConnection = 'No se pudo conectar al servidor :(';